Default to local overlay_filter
Up until Pandoc 1.19 the --filter option only looked in $PATH. To use a
local file it had to be explicitly specified as one. I want to default
to the local file because 1) I can control it, and 2) I don't want to
pollute the host system. I don't have access to Pandoc 1.19 so I have to
change the path.