Install Ubuntu dependencies with a cannon
Running this script installs enough dependencies to successfully 'make
all' except for missing fonts not included in standard packages --
problematic for the examples, which use a non-standard (proprietary)
font, but otherwise good. The script probably installs more dependencies
than strictly necessary, but not many, and that's quite typical of tex
packages anyway.